+++ data-plane/src/server/tls/cert_resolver.rs
@@ -403,12 +403,7 @@ mod tests {
         let team_uuid = "team_456".to_string();
         let cage_uuid = "cage_123".to_string();
         let cage_name = "a-cage".to_string();
-        let ctx = CageContext::new(
-            app_uuid,
-            team_uuid,
-            cage_uuid,
-            cage_name
-        );
+        let ctx = CageContext::new(app_uuid, team_uuid, cage_uuid, cage_name);
         let hostname = ctx.get_cert_name();
         assert!(hostname.ends_with(".dev"));
     }
