+++ e2e-tests/scripts/start-data-plane.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 
-echo {\"apiKeyAuth\":\"true\",\"egress\":{},\"trxLoggingEnabled\":\"true\"} > /etc/dataplane-config.json
+echo {\"api_key_auth\":true,\"egress\":{\"ports\": \"443\", \"allow_list\": \"jsonplaceholder.typicode.com\"},\"trx_logging_enabled\":true} > /etc/dataplane-config.json
 
 exec $DATA_PLANE_EXECUTABLE_PATH
